requirements.txt
streamlit
pytesseract
Pillow
PyMuPDF